# perf(*) reuse cdata to improve performance