Although still slower than ngx.md5_bin, we are around 15% faster than previous self.

```
resty.openssl.digest(md5) took 1870 ms
resty.openssl.digest(md5,old) took 2159 ms
resty.md5 took 2423 ms
ngx.md5_bin took 1341 ms

resty.openssl.cipher(aes,encrypt) took 598 ms
resty.openssl.cipher(aes,encrypt,old) took 818ms
resty.aes took 1891 ms
```

Because EVP contextes are reused, we are still way faster than resty.string implementation.

<details><summary> digest </summary>

```lua
local md5 = ngx.md5_bin
local resty_md5 = require "resty.md5"

local function runperf()
  local digest = require "resty.openssl.digest"
  local update_time = ngx.update_time
  local now = ngx.now

  local md5_bin = digest.new("md5")


  local m = "my:cache:key:b864ded6-f344-49f7-b999-dd9f6aaaad7f"

  update_time()
  local s = now()*1000
  for _i= 1, 10000000 do
    md5_bin:final(m)
    md5_bin:reset()
  end
  update_time()
  local e = now()*1000

  ngx.log(ngx.ERR, "resty.openssl.digest(md5) took ", e - s, " ms")

  update_time()
  local s = now()*1000
  for _i= 1, 10000000 do
      local md5 = resty_md5:new()
      md5:update(m)
      md5:final()
  end
  update_time()
  local e = now()*1000

  ngx.log(ngx.ERR, "resty.md5 took ", e - s, " ms")

  update_time()
  s = now()*1000
  for _i= 1, 10000000 do
    ngx.md5_bin(m)
  end
  update_time()
  e = now()*1000

  ngx.log(ngx.ERR, "ngx.md5 took ", e - s, " ms")

end
runperf()
```
</details>

<details><summary> cipher </summary>

```lua
local aes = require "resty.aes"

local function runperf()
  local cipher = require "resty.openssl.cipher"
  local update_time = ngx.update_time
  local now = ngx.now

  local aes_128 = cipher.new("aes-128-cbc")
  local key = string.rep("0", 16)
  local iv = string.rep("0", 16)

  local m = "my:cache:key:b864ded6-f344-49f7-b999-dd9f6aaaad7f"

  local d
  update_time()
  local s = now()*1000
  for _i= 1, 1000000 do
      aes_128:init(key, iv, {
          is_encrypt = true,
      })
      d = aes_128:final(m)
  end
  update_time()
  local e = now()*1000

  ngx.log(ngx.ERR, "resty.openssl.cipher(aes,encrypt) ", e - s, " ms")
  print(d)

  local _hash = { iv = iv }
  update_time()
  local s = now()*1000
  for _i= 1, 1000000 do
      local aes_128_cbc_md5 = aes:new(key, nil, nil, _hash )
      d = aes_128_cbc_md5:encrypt(m)
  end
  update_time()
  local e = now()*1000

  ngx.log(ngx.ERR, "resty.aes(encrypt) took ", e - s, " ms")
  print(d)

  update_time()
  local s = now()*1000
  for _i= 1, 1000000 do
    d = aes_128:encrypt(key, iv, m)
  end
  update_time()
  local e = now()*1000

  ngx.log(ngx.ERR, "resty.openssl.digest oneshot took ", e - s, " ms")
  print(d)
end
runperf()
```

</details>